move xs stuff to util
